Updated the civo cloud provider
