+++ internal/cloudproviders/civo/civo_test.go
@@ -1,67 +1,18 @@
 package civo
 
 import (
-	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"os"
-	"path/filepath"
 	"strings"
 	"testing"
 
-	storageTypes "github.com/ksctl/ksctl/pkg/types/storage"
-
 	"github.com/civo/civogo"
-	"github.com/ksctl/ksctl/pkg/logger"
-	"github.com/ksctl/ksctl/pkg/types/controllers/cloud"
-
-	localstate "github.com/ksctl/ksctl/internal/storage/local"
 	"github.com/ksctl/ksctl/pkg/helpers"
 	"github.com/ksctl/ksctl/pkg/helpers/consts"
 	ksctlErrors "github.com/ksctl/ksctl/pkg/helpers/errors"
-	"github.com/ksctl/ksctl/pkg/types"
 	"gotest.tools/v3/assert"
 )
-
-var (
-	fakeClientHA *CivoProvider
-	storeHA      types.StorageFactory
-
-	fakeClientManaged *CivoProvider
-	storeManaged      types.StorageFactory
-
-	fakeClientVars *CivoProvider
-	storeVars      types.StorageFactory
-
-	dir          = filepath.Join(os.TempDir(), "ksctl-civo-test")
-	parentCtx    context.Context
-	parentLogger types.LoggerFactory = logger.NewStructuredLogger(-1, os.Stdout)
-)
-
-func TestMain(m *testing.M) {
-	parentCtx = context.WithValue(context.TODO(), consts.KsctlCustomDirLoc, dir)
-
-	fakeClientVars, _ = NewClient(parentCtx, types.Metadata{
-		ClusterName: "demo",
-		Region:      "LON1",
-		Provider:    consts.CloudCivo,
-		IsHA:        true,
-	}, parentLogger, &storageTypes.StorageDocument{}, ProvideClient)
-
-	storeVars = localstate.NewClient(parentCtx, parentLogger)
-	_ = storeVars.Setup(consts.CloudCivo, "LON1", "demo", consts.ClusterTypeHa)
-	_ = storeVars.Connect()
-
-	exitVal := m.Run()
-
-	fmt.Println("Cleanup..")
-	if err := os.RemoveAll(dir); err != nil {
-		panic(err)
-	}
-
-	os.Exit(exitVal)
-}
 
 func TestIsValidK8sVersion(t *testing.T) {
 	ver, _ := fakeClientVars.client.ListAvailableKubernetesVersions()
@@ -408,508 +359,4 @@ func TestDeleteVarCluster(t *testing.T) {
 	if err := storeVars.DeleteCluster(); err != nil {
 		t.Fatal(err)
 	}
-}
-
-func checkCurrentStateFile(t *testing.T) {
-
-	if err := storeManaged.Setup(consts.CloudCivo, mainStateDocument.Region, mainStateDocument.ClusterName, consts.ClusterTypeMang); err != nil {
-		t.Fatal(err)
-	}
-	read, err := storeManaged.Read()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	assert.DeepEqual(t, mainStateDocument, read)
-}
-
-func checkCurrentStateFileHA(t *testing.T) {
-
-	if err := storeHA.Setup(consts.CloudCivo, mainStateDocument.Region, mainStateDocument.ClusterName, consts.ClusterTypeHa); err != nil {
-		t.Fatal(err)
-	}
-	read, err := storeHA.Read()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	assert.DeepEqual(t, mainStateDocument, read)
-}
-
-func TestManagedCluster(t *testing.T) {
-
-	func() {
-		fakeClientManaged, _ = NewClient(parentCtx, types.Metadata{
-			ClusterName: "demo-managed",
-			Region:      "LON1",
-			Provider:    consts.CloudCivo,
-		}, parentLogger, &storageTypes.StorageDocument{}, ProvideClient)
-
-		storeManaged = localstate.NewClient(parentCtx, parentLogger)
-		_ = storeManaged.Setup(consts.CloudCivo, "LON1", "demo-managed", consts.ClusterTypeMang)
-		_ = storeManaged.Connect()
-	}()
-
-	t.Run("init state", func(t *testing.T) {
-
-		if err := fakeClientManaged.InitState(storeManaged, consts.OperationCreate); err != nil {
-			t.Fatalf("Unable to init the state for fresh start, Reason: %v", err)
-		}
-
-		assert.Equal(t, clusterType, consts.ClusterTypeMang, "clustertype should be managed")
-		assert.Equal(t, mainStateDocument.CloudInfra.Civo.B.IsCompleted, false, "cluster should not be completed")
-
-		_, err := storeManaged.Read()
-		if err == nil {
-			t.Fatalf("State file and cluster directory present where it should not be")
-		}
-	})
-
-	t.Run("Create network", func(t *testing.T) {
-		assert.Equal(t, fakeClientManaged.Name("fake-net").NewNetwork(storeManaged), nil, "Network should be created")
-		assert.Equal(t, mainStateDocument.CloudInfra.Civo.B.IsCompleted, false, "cluster should not be completed")
-		assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.NetworkID) > 0, "network id not saved")
-
-		checkCurrentStateFile(t)
-	})
-
-	t.Run("Create managed cluster", func(t *testing.T) {
-
-		fakeClientManaged.CNI("cilium")
-		fakeClientManaged.Application([]string{"abcd"})
-
-		assert.Equal(t, fakeClientManaged.Name("fake").VMType("g4s.kube.small").NewManagedCluster(storeManaged, 5), nil, "managed cluster should be created")
-
-		assert.Equal(t, mainStateDocument.CloudInfra.Civo.B.IsCompleted, true, "cluster should not be completed")
-
-		assert.Equal(t, mainStateDocument.CloudInfra.Civo.NoManagedNodes, 5)
-		assert.Equal(t, mainStateDocument.CloudInfra.Civo.B.KubernetesVer, fakeClientManaged.metadata.k8sVersion)
-		assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.ManagedClusterID) > 0, "Managed clusterID not saved")
-
-		_, err := storeManaged.Read()
-		if err != nil {
-			t.Fatalf("kubeconfig should not be absent")
-		}
-		checkCurrentStateFile(t)
-	})
-
-	t.Run("Get cluster managed", func(t *testing.T) {
-		expected := []cloud.AllClusterData{
-			{
-				Name:          fakeClientManaged.clusterName,
-				CloudProvider: consts.CloudCivo,
-				ClusterType:   consts.ClusterTypeMang,
-				NetworkID:     "fake-net",
-				ManagedK8sID:  "fake-k8s",
-				Region:        fakeClientManaged.region,
-				NoMgt:         mainStateDocument.CloudInfra.Civo.NoManagedNodes,
-				Mgt:           cloud.VMData{VMSize: "g4s.kube.small"},
-
-				K8sDistro:  "managed",
-				K8sVersion: mainStateDocument.CloudInfra.Civo.B.KubernetesVer,
-			},
-		}
-		got, err := fakeClientManaged.GetRAWClusterInfos(storeManaged)
-		assert.NilError(t, err, "no error should be there")
-		assert.DeepEqual(t, got, expected)
-	})
-
-	t.Run("Delete managed cluster", func(t *testing.T) {
-		assert.Equal(t, fakeClientManaged.DelManagedCluster(storeManaged), nil, "managed cluster should be deleted")
-
-		assert.Equal(t, len(mainStateDocument.CloudInfra.Civo.ManagedClusterID), 0, "managed cluster id still present")
-
-		checkCurrentStateFile(t)
-	})
-
-	t.Run("Delete Network cluster", func(t *testing.T) {
-		assert.Equal(t, fakeClientManaged.DelNetwork(storeManaged), nil, "Network should be deleted")
-
-		assert.Equal(t, len(mainStateDocument.CloudInfra.Civo.NetworkID), 0, "network id still present")
-		// at this moment the file is not present
-		_, err := storeManaged.Read()
-		if err == nil {
-			t.Fatalf("State file and cluster directory still present")
-		}
-	})
-
-}
-
-func TestHACluster(t *testing.T) {
-	fakeClientHA, _ = NewClient(parentCtx, types.Metadata{
-		ClusterName: "demo-ha",
-		Region:      "LON1",
-		Provider:    consts.CloudCivo,
-		IsHA:        true,
-		NoCP:        7,
-		NoDS:        5,
-		NoWP:        10,
-		K8sDistro:   consts.K8sK3s,
-	}, parentLogger, &storageTypes.StorageDocument{}, ProvideClient)
-
-	storeHA = localstate.NewClient(parentCtx, parentLogger)
-	_ = storeHA.Setup(consts.CloudCivo, "LON1", "demo-ha", consts.ClusterTypeHa)
-	_ = storeHA.Connect()
-
-	fakeClientHA.metadata.noCP = 7
-	fakeClientHA.metadata.noDS = 5
-	fakeClientHA.metadata.noWP = 10
-
-	t.Run("init state", func(t *testing.T) {
-
-		if err := fakeClientHA.InitState(storeHA, consts.OperationCreate); err != nil {
-			t.Fatalf("Unable to init the state for fresh start, Reason: %v", err)
-		}
-
-		assert.Equal(t, clusterType, consts.ClusterTypeHa, "clustertype should be managed")
-		assert.Equal(t, mainStateDocument.CloudInfra.Civo.B.IsCompleted, false, "cluster should not be completed")
-
-		_, err := storeHA.Read()
-		if err == nil {
-			t.Fatalf("State file and cluster directory present where it should not be")
-		}
-	})
-
-	t.Run("Create network", func(t *testing.T) {
-		assert.Equal(t, fakeClientHA.Name("fake-net").NewNetwork(storeHA), nil, "Network should be created")
-		assert.Equal(t, mainStateDocument.CloudInfra.Civo.B.IsCompleted, false, "cluster should not be completed")
-		assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.NetworkID) > 0, "network id not saved")
-
-		checkCurrentStateFileHA(t)
-	})
-
-	t.Run("Create ssh", func(t *testing.T) {
-
-		assert.Equal(t, fakeClientHA.Name("fake-ssh").CreateUploadSSHKeyPair(storeHA), nil, "ssh key failed")
-
-		assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.B.SSHID) > 0, "sshid must be present")
-		assert.Equal(t, mainStateDocument.CloudInfra.Civo.B.SSHUser, "root", "ssh user not set")
-
-		assert.Equal(t, mainStateDocument.CloudInfra.Civo.B.IsCompleted, false, "cluster should not be completed")
-		checkCurrentStateFileHA(t)
-	})
-
-	t.Run("Create Firewalls", func(t *testing.T) {
-
-		t.Run("Controlplane", func(t *testing.T) {
-			fakeClientHA.Role(consts.RoleCp)
-			fakeClientHA.Name("fake-cp")
-
-			assert.Equal(t, fakeClientHA.NewFirewall(storeHA), nil, "new firewall failed")
-
-			assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.FirewallIDControlPlanes) > 0, "firewallID for controlplane absent")
-		})
-		t.Run("Workerplane", func(t *testing.T) {
-			fakeClientHA.Role(consts.RoleWp)
-			fakeClientHA.Name("fake-wp")
-
-			assert.Equal(t, fakeClientHA.NewFirewall(storeHA), nil, "new firewall failed")
-			assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.FirewallIDWorkerNodes) > 0, "firewallID for workerplane absent")
-		})
-		t.Run("Loadbalancer", func(t *testing.T) {
-			fakeClientHA.Role(consts.RoleLb)
-			fakeClientHA.Name("fake-lb")
-
-			assert.Equal(t, fakeClientHA.NewFirewall(storeHA), nil, "new firewall failed")
-			assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.FirewallIDLoadBalancer) > 0, "firewallID for loadbalancer absent")
-		})
-		t.Run("Datastore", func(t *testing.T) {
-			fakeClientHA.Role(consts.RoleDs)
-			fakeClientHA.Name("fake-ds")
-
-			assert.Equal(t, fakeClientHA.NewFirewall(storeHA), nil, "new firewall failed")
-			assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.FirewallIDDatabaseNodes) > 0, "firewallID for datastore absent")
-		})
-
-		checkCurrentStateFileHA(t)
-	})
-
-	t.Run("Create VMs", func(t *testing.T) {
-		t.Run("Loadbalancer", func(t *testing.T) {
-			fakeClientHA.Role(consts.RoleLb)
-			fakeClientHA.Name("fake-lb-vm")
-			fakeClientHA.VMType("g4s.kube.small")
-
-			assert.Equal(t, fakeClientHA.NewVM(storeHA, 0), nil, "new vm failed")
-			assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoLoadBalancer.VMID) > 0, "loadbalancer VM id absent")
-
-			assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoLoadBalancer.PublicIP) > 0, "loadbalancer ipv4 absent")
-			assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoLoadBalancer.PrivateIP) > 0, "loadbalancer private ipv4 absent")
-			assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoLoadBalancer.HostName) > 0, "loadbalancer hostname absent")
-
-			checkCurrentStateFileHA(t)
-		})
-		t.Run("Controlplanes", func(t *testing.T) {
-
-			if _, err := fakeClientHA.NoOfControlPlane(fakeClientHA.metadata.noCP, true); err != nil {
-				t.Fatalf("Failed to set the controlplane, err: %v", err)
-			}
-
-			for i := 0; i < fakeClientHA.metadata.noCP; i++ {
-				t.Run("controlplane", func(t *testing.T) {
-
-					fakeClientHA.Name(fmt.Sprintf("fake-cp-%d", i))
-					fakeClientHA.VMType("g4s.kube.small")
-					fakeClientHA.Role(consts.RoleCp)
-
-					assert.Equal(t, fakeClientHA.NewVM(storeHA, i), nil, "new vm failed")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoControlPlanes.VMIDs[i]) > 0, "controlplane VM id absent")
-
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoControlPlanes.PublicIPs[i]) > 0, "controlplane ipv4 absent")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoControlPlanes.PrivateIPs[i]) > 0, "controlplane private ipv4 absent")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoControlPlanes.Hostnames[i]) > 0, "controlplane hostname absent")
-
-					checkCurrentStateFileHA(t)
-				})
-			}
-		})
-
-		t.Run("Datastores", func(t *testing.T) {
-
-			if _, err := fakeClientHA.NoOfDataStore(fakeClientHA.metadata.noDS, true); err != nil {
-				t.Fatalf("Failed to set the datastore")
-			}
-
-			for i := 0; i < fakeClientHA.metadata.noDS; i++ {
-				t.Run("datastore", func(t *testing.T) {
-
-					fakeClientHA.Name(fmt.Sprintf("fake-ds-%d", i))
-					fakeClientHA.VMType("g4s.kube.small")
-					fakeClientHA.Role(consts.RoleDs)
-
-					assert.Equal(t, fakeClientHA.NewVM(storeHA, i), nil, "new vm failed")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoDatabase.VMIDs[i]) > 0, "datastore VM id absent")
-
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoDatabase.PublicIPs[i]) > 0, "datastore ipv4 absent")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoDatabase.PrivateIPs[i]) > 0, "datastore private ipv4 absent")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoDatabase.Hostnames[i]) > 0, "datastore hostname absent")
-
-					checkCurrentStateFileHA(t)
-				})
-			}
-		})
-		t.Run("Workplanes", func(t *testing.T) {
-
-			if _, err := fakeClientHA.NoOfWorkerPlane(storeHA, fakeClientHA.metadata.noWP, true); err != nil {
-				t.Fatalf("Failed to set the workerplane")
-			}
-
-			for i := 0; i < fakeClientHA.metadata.noWP; i++ {
-				t.Run("workerplane", func(t *testing.T) {
-
-					fakeClientHA.Name(fmt.Sprintf("fake-wp-%d", i))
-					fakeClientHA.Role(consts.RoleWp)
-					fakeClientHA.VMType("g4s.kube.small")
-
-					assert.Equal(t, fakeClientHA.NewVM(storeHA, i), nil, "new vm failed")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoWorkerPlanes.VMIDs[i]) > 0, "workerplane VM id absent")
-
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoWorkerPlanes.PublicIPs[i]) > 0, "workerplane ipv4 absent")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoWorkerPlanes.PrivateIPs[i]) > 0, "workerplane private ipv4 absent")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoWorkerPlanes.Hostnames[i]) > 0, "workerplane hostname absent")
-
-					checkCurrentStateFileHA(t)
-				})
-			}
-
-			assert.Equal(t, mainStateDocument.CloudInfra.Civo.B.IsCompleted, true, "cluster should be completed")
-		})
-	})
-
-	fmt.Println(fakeClientHA.GetHostNameAllWorkerNode())
-	t.Run("get hostname of workerplanes", func(t *testing.T) {
-		expected := mainStateDocument.CloudInfra.Civo.InfoWorkerPlanes.Hostnames
-
-		got := fakeClientHA.GetHostNameAllWorkerNode()
-		assert.DeepEqual(t, got, expected)
-	})
-
-	t.Run("check getState()", func(t *testing.T) {
-		expected, err := fakeClientHA.GetStateFile(storeHA)
-		assert.NilError(t, err, "no error should be there for getstate")
-
-		got, _ := json.Marshal(mainStateDocument)
-		assert.DeepEqual(t, string(got), expected)
-	})
-
-	t.Run("Get cluster ha", func(t *testing.T) {
-		_e := cloud.VMData{
-			VMSize:     "g4s.kube.small",
-			VMID:       "vm-fake",
-			FirewallID: "fake-fw",
-			PublicIP:   "A.B.C.D",
-			PrivateIP:  "192.168.1.2",
-		}
-		expected := []cloud.AllClusterData{
-			{
-				Name:          fakeClientHA.clusterName,
-				CloudProvider: consts.CloudCivo,
-				ClusterType:   consts.ClusterTypeHa,
-				Region:        fakeClientHA.region,
-				NoWP:          fakeClientHA.noWP,
-				NoCP:          fakeClientHA.noCP,
-				NoDS:          fakeClientHA.noDS,
-				SSHKeyID:      "fake-ssh",
-				NetworkID:     "fake-net",
-				LB:            _e,
-				WP: []cloud.VMData{
-					_e, _e, _e, _e,
-					_e, _e, _e, _e,
-					_e, _e,
-				},
-				CP: []cloud.VMData{
-					_e, _e, _e, _e,
-					_e, _e, _e,
-				},
-				DS: []cloud.VMData{
-					_e, _e, _e, _e,
-					_e,
-				},
-				K8sDistro:  "",
-				K8sVersion: mainStateDocument.CloudInfra.Civo.B.KubernetesVer,
-			},
-		}
-		got, err := fakeClientHA.GetRAWClusterInfos(storeHA)
-		assert.NilError(t, err, "no error should be there")
-		assert.DeepEqual(t, got, expected)
-	})
-
-	// explicit clean
-	mainStateDocument = &storageTypes.StorageDocument{}
-
-	// use init state firest
-	t.Run("init state deletion", func(t *testing.T) {
-
-		if err := fakeClientHA.InitState(storeHA, consts.OperationDelete); err != nil {
-			t.Fatalf("Unable to init the state for delete, Reason: %v", err)
-		}
-
-		assert.Equal(t, clusterType, consts.ClusterTypeHa, "clustertype should be managed")
-	})
-
-	t.Run("Get all counters", func(t *testing.T) {
-		var err error
-		fakeClientHA.metadata.noCP, err = fakeClientHA.NoOfControlPlane(-1, false)
-		assert.Assert(t, err == nil)
-
-		fakeClientHA.metadata.noWP, err = fakeClientHA.NoOfWorkerPlane(storeHA, -1, false)
-		assert.Assert(t, err == nil)
-
-		fakeClientHA.metadata.noDS, err = fakeClientHA.NoOfDataStore(-1, false)
-		assert.Assert(t, err == nil)
-	})
-
-	t.Run("Delete VMs", func(t *testing.T) {
-		t.Run("Loadbalancer", func(t *testing.T) {
-			fakeClientHA.Role(consts.RoleLb)
-
-			assert.Equal(t, fakeClientHA.DelVM(storeHA, 0), nil, "del vm failed")
-			assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoLoadBalancer.VMID) == 0, "loadbalancer VM id absent")
-
-			assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoLoadBalancer.PublicIP) == 0, "loadbalancer ipv4 absent")
-			assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoLoadBalancer.PrivateIP) == 0, "loadbalancer private ipv4 present")
-			assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoLoadBalancer.HostName) == 0, "loadbalancer hostname present")
-
-			checkCurrentStateFileHA(t)
-		})
-
-		t.Run("Workerplane", func(t *testing.T) {
-
-			for i := 0; i < fakeClientHA.metadata.noWP; i++ {
-				t.Run("workerplane", func(t *testing.T) {
-					fakeClientHA.Role(consts.RoleWp)
-
-					assert.Equal(t, fakeClientHA.DelVM(storeHA, i), nil, "del vm failed")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoWorkerPlanes.VMIDs[i]) == 0, "workerplane VM id present")
-
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoWorkerPlanes.PublicIPs[i]) == 0, "workerplane ipv4 present")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoWorkerPlanes.PrivateIPs[i]) == 0, "workerplane private ipv4 present")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoWorkerPlanes.Hostnames[i]) == 0, "workerplane hostname present")
-
-					checkCurrentStateFileHA(t)
-				})
-			}
-		})
-		t.Run("Controlplane", func(t *testing.T) {
-
-			for i := 0; i < fakeClientHA.metadata.noCP; i++ {
-				t.Run("controlplane", func(t *testing.T) {
-					fakeClientHA.Role(consts.RoleCp)
-
-					assert.Equal(t, fakeClientHA.DelVM(storeHA, i), nil, "del vm failed")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoControlPlanes.VMIDs[i]) == 0, "controlplane VM id present")
-
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoControlPlanes.PublicIPs[i]) == 0, "controlplane ipv4 present")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoControlPlanes.PrivateIPs[i]) == 0, "controlplane private ipv4 present")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoControlPlanes.Hostnames[i]) == 0, "controlplane hostname present")
-
-					checkCurrentStateFileHA(t)
-				})
-			}
-		})
-		t.Run("DataStore", func(t *testing.T) {
-
-			for i := 0; i < fakeClientHA.metadata.noDS; i++ {
-				t.Run("datastore", func(t *testing.T) {
-					fakeClientHA.Role(consts.RoleDs)
-
-					assert.Equal(t, fakeClientHA.DelVM(storeHA, i), nil, "del vm failed")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoDatabase.VMIDs[i]) == 0, "datastore VM id present")
-
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoDatabase.PublicIPs[i]) == 0, "datastore ipv4 present")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoDatabase.PrivateIPs[i]) == 0, "datastore private ipv4 present")
-					assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.InfoDatabase.Hostnames[i]) == 0, "datastore hostname present")
-
-					checkCurrentStateFileHA(t)
-				})
-			}
-		})
-	})
-
-	t.Run("Delete Firewalls", func(t *testing.T) {
-
-		t.Run("Controlplane", func(t *testing.T) {
-			fakeClientHA.Role(consts.RoleCp)
-
-			assert.Equal(t, fakeClientHA.DelFirewall(storeHA), nil, "del firewall failed")
-
-			assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.FirewallIDControlPlanes) == 0, "firewallID for controlplane present")
-		})
-		t.Run("Workerplane", func(t *testing.T) {
-			fakeClientHA.Role(consts.RoleWp)
-
-			assert.Equal(t, fakeClientHA.DelFirewall(storeHA), nil, "new firewall failed")
-			assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.FirewallIDWorkerNodes) == 0, "firewallID for workerplane present")
-		})
-		t.Run("Loadbalancer", func(t *testing.T) {
-			fakeClientHA.Role(consts.RoleLb)
-
-			assert.Equal(t, fakeClientHA.DelFirewall(storeHA), nil, "new firewall failed")
-			assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.FirewallIDLoadBalancer) == 0, "firewallID for loadbalancer present")
-		})
-		t.Run("Datastore", func(t *testing.T) {
-			fakeClientHA.Role(consts.RoleDs)
-
-			assert.Equal(t, fakeClientHA.DelFirewall(storeHA), nil, "new firewall failed")
-			assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.FirewallIDDatabaseNodes) == 0, "firewallID for datastore present")
-		})
-
-		checkCurrentStateFileHA(t)
-	})
-
-	t.Run("Delete ssh", func(t *testing.T) {
-
-		assert.Equal(t, fakeClientHA.DelSSHKeyPair(storeHA), nil, "ssh key failed")
-
-		assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.B.SSHID) == 0, "sshid still present")
-		assert.Equal(t, mainStateDocument.CloudInfra.Civo.B.SSHUser, "", "ssh user set")
-
-		checkCurrentStateFileHA(t)
-	})
-
-	t.Run("Delete network", func(t *testing.T) {
-		assert.Equal(t, fakeClientHA.DelNetwork(storeHA), nil, "Network should be deleted")
-		assert.Assert(t, len(mainStateDocument.CloudInfra.Civo.NetworkID) == 0, "network id still present")
-	})
-
 }
